+++ inst/www/js/handsontable-selectize-editor.js
@@ -22,10 +22,7 @@
   SelectizeEditor.prototype.createElements = function() {
     this.$body = $(document.body);
 
-    this.TEXTAREA = document.createElement("select");
-    // fix for handsontable >=5.0.0
-    this.TEXTAREA.select = function() {};
-    this.$select = $(this.TEXTAREA);
+    this.TEXTAREA = document.createElement("TEXTAREA");
 
     Handsontable.dom.addClass(this.TEXTAREA, "handsontableInput");
 
@@ -43,17 +40,11 @@
 
     this.TEXTAREA_PARENT.appendChild(this.TEXTAREA);
 
-    this.instance.rootElement.appendChild(this.TEXTAREA_PARENT);
-
-    // possible fix for copy/paste
-    /*this.TEXTAREA = document.createElement("TEXTAREA");
     var select = document.createElement("SELECT");
     this.TEXTAREA_PARENT.appendChild(select);
     this.$select = $(select);
 
-    this.TEXTAREA.select = function() {
-      this.focus();
-    };*/
+    this.instance.rootElement.appendChild(this.TEXTAREA_PARENT);
 
     var that = this;
     this.instance._registerTimeout(
@@ -82,8 +73,6 @@
     var ctrlDown = (event.ctrlKey || event.metaKey) && !event.altKey;
     var sel = editor.selectize;
     var multiple = editor && sel.settings.mode == "multi";
-console.log(ctrlDown);
-console.log( event.keyCode);
     // Process only events that have been fired in the editor
     if (event.target.tagName !== "INPUT") {
       // except BACKSPACEs for deleting the last item since there won't be an INPUT if selectize is full